Add the user_id in the filter
If you don't put user_id, this default value will be the id of administrator
and the others users cannot use the filter.
We get an Acl error saying: 'Access denied for ir_filters models'.
Definition of this field:
user_id': fields.many2one('res.users', 'User', ondelete='cascade',
help='The user this filter is private to. When left empty the filter is public'
'and available to all users.)